+++ apps/web/app/[project-id]/data/[table-id]/page.tsx
@@ -1,7 +1,7 @@
 "use client";
 
 import { useParams } from "next/navigation";
-import { TableEditor } from "../TableEditor";
+import { TableEditor } from "../components/TableEditor";
 
 export default function Page() {
    const params = useParams();
