+++ apps/web/app/[project-id]/data/components/TableEditor.tsx
@@ -1,17 +1,14 @@
 "use client";
 
-import {
-   ColumnHelper,
-   createColumnHelper,
-   flexRender,
-   getCoreRowModel,
-   useReactTable,
-   ColumnResizeMode,
-} from "@tanstack/react-table";
+import { ColumnHelper } from "@tanstack/react-table";
 import { useRouter } from "next/navigation";
-import { Reducer, useEffect, useMemo, useReducer, useState } from "react";
+import { Reducer, useEffect, useReducer, useState } from "react";
 import useSWR from "swr";
+import { ColumnConfigMenu } from "./ColumnConfigMenu";
+import { RowConfigMenu } from "./RowConfigMenu";
 import "./style.css";
+import { TableCanva } from "./TableCanva";
+import { TableEditorSidebar } from "./TableEditorSidebar";
 
 const mockApiBuilder = (projectId: string) => {
    const base = process.env["NEXT_PUBLIC_BASE_URL"];
@@ -96,7 +93,7 @@ type TableEditorProps = {
    tableId: string;
 };
 
-type TableCanvaProps = {
+export type TableCanvaProps = {
    state: CanvaState;
    dispatch: any;
    setRowConfig: any;
@@ -120,7 +117,12 @@ const canvaReducer: Reducer<CanvaState, CanvaAction> = (state, action) => {
       case "remove-row":
          return { ...state };
       case "select-table": // change table
-         return { ...state, tableId: action.tableId };
+         return {
+            ...state,
+            tableId: action.tableId,
+            isColumnConfigOpen: false,
+            isRowConfigOpen: false,
+         };
       case "set-data":
          return { ...state, data: action.payload };
       default:
@@ -184,96 +186,6 @@ const useTableEditor = (projectId, tableId) => {
    return [canvaState, canvaDispatch, configState, configDispatch] as const;
 };
 
-export function EditorConfigMenu({ isOpen, onClose }: EditorConfigMenuProps) {
-   const [configItems, setConfigItems] = useState<EditorConfigMenuItem[]>([]);
-
-   const toggleConfigMenu = () => {
-      onClose();
-   };
-
-   return (
-      <div className={`config-menu ${isOpen ? "open" : ""}`}>
-         <button onClick={toggleConfigMenu}>
-            {isOpen ? "Close Config Menu" : "Open Config Menu"}
-         </button>
-      </div>
-   );
-}
-
-export function RowConfigMenu({ isOpen, onClose, dispatch }) {
-   return (
-      <div className={`row-config-menu ${isOpen ? "open" : ""}`}>
-         <button onClick={() => dispatch({ type: "open-config-insert-row" })}>
-            Insert Row
-         </button>
-      </div>
-   );
-}
-
-export function ColumnConfigMenu({ isOpen, onClose, dispatch }) {
-   const [columnName, setColumnName] = useState("");
-   const [dataType, setDataType] = useState("Text");
-   const [defaultValue, setDefaultValue] = useState("");
-
-   const handleAddColumn = () => {
-      dispatch({
-         type: "open-config-add-column",
-         payload: { columnName, dataType, defaultValue },
-      });
-      onClose();
-   };
-   return (
-      <div className={`column-config-menu ${isOpen ? "open" : ""}`}>
-         <button onClick={handleAddColumn}>Add Column</button>
-         <button onClick={onClose}>Close</button>
-         <div className={`column-options`}>
-            <div className={`column-option`}>
-               <div className={`column-option-label`}>Column Name</div>
-               <input
-                  type="text"
-                  value={columnName}
-                  onChange={(e) => setColumnName(e.target.value)}
-               />
-
-               <div className={`column-option-label`}>Data Type</div>
-               <select
-                  value={dataType}
-                  onChange={(e) => setDataType(e.target.value)}
-               >
-                  <option>Text</option>
-                  <option>Number</option>
-                  <option>Date</option>
-                  <option>Boolean</option>
-                  <option>Image</option>
-                  <option>File</option>
-               </select>
-
-               <div className={`column-option-label`}>Default Value</div>
-               <input
-                  type="text"
-                  value={defaultValue}
-                  onChange={(e) => setDefaultValue(e.target.value)}
-               />
-            </div>
-         </div>
-      </div>
-   );
-}
-
-export function TableCanva({
-   state,
-   dispatch,
-   setRowConfig,
-   setColumnConfig,
-}: TableCanvaProps) {
-   return (
-      <div className="canva">
-         <CanvaToolbar />
-         <TableRenderer data={state.data} dispatch={dispatch} />
-      </div>
-   );
-}
-
 const FilterToolbarItem = () => (
    <div className="filter-toolbar-item" style={{ margin: "10px" }}>
       <Dropdown
@@ -332,7 +244,7 @@ const Dropdown = ({ title, options, onSelect }) => {
 const DropdownOption = ({ value, children, onSelect }) => (
    <div onClick={() => onSelect(value)}>{children}</div>
 );
-function CanvaToolbar() {
+export function CanvaToolbar() {
    const [selectedTool, setSelectedTool] = useState(null);
 
    return (
@@ -360,7 +272,7 @@ type ColumnProps = {
    children?: ColumnProps[];
 };
 
-const parseColumns = (
+export const parseColumns = (
    columnsHelper: ColumnHelper<unknown>,
    columns: ColumnProps[],
    onAddColumn?: any
@@ -394,124 +306,6 @@ const parseColumns = (
    return [...cols, addCol];
 };
 
-function TableRenderer({ data, dispatch }) {
-   const columnHelper = useMemo(() => createColumnHelper(), []);
-   const [columnResizeMode, setColumnResizeMode] =
-      useState<ColumnResizeMode>("onChange");
-
-   const columns = useMemo(() => {
-      return parseColumns(columnHelper, data?.columns ?? [], dispatch);
-   }, [data]);
-
-   const rows = useMemo(() => {
-      return data?.rows ?? [];
-   }, [data]);
-
-   const coreRowModel = getCoreRowModel();
-
-   const table = useReactTable({
-      data: rows,
-      columns,
-      columnResizeMode,
-      getCoreRowModel: coreRowModel,
-   });
-
-   if (!data) {
-      return <div>Loading...</div>;
-   }
-
-   return (
-      <div
-         className={"table"}
-         style={{
-            width: "100vh",
-            height: "100%",
-            overflowX: "scroll",
-            overflowY: "auto",
-         }}
-      >
-         <table
-            {...{
-               style: {
-                  width: table.getCenterTotalSize(),
-               },
-            }}
-         >
-            <thead>
-               {table.getHeaderGroups().map((headerGroup) => (
-                  <tr key={headerGroup.id}>
-                     {headerGroup.headers.map((header) => (
-                        <>
-                           <th
-                              key={header.id}
-                              colSpan={header.colSpan}
-                              style={{
-                                 border: "1px solid black",
-                                 width: header.getSize(),
-                              }}
-                           >
-                              {header.isPlaceholder ? null : (
-                                 <>
-                                    <div>
-                                       {flexRender(
-                                          header.column.columnDef.header,
-                                          header.getContext()
-                                       )}
-                                    </div>
-                                 </>
-                              )}
-
-                              <div
-                                 onMouseDown={header.getResizeHandler()}
-                                 onTouchStart={header.getResizeHandler()}
-                                 className={`resizer ${
-                                    header.column.getIsResizing()
-                                       ? "isResizing"
-                                       : ""
-                                 }`}
-                                 style={{
-                                    transform:
-                                       columnResizeMode === "onEnd" &&
-                                       header.column.getIsResizing()
-                                          ? `translateX(${
-                                               table.getState().columnSizingInfo
-                                                  .deltaOffset
-                                            }px)`
-                                          : "",
-                                 }}
-                              />
-                           </th>
-                        </>
-                     ))}
-                  </tr>
-               ))}
-            </thead>
-            <tbody>
-               {table.getRowModel().rows.map((row, index) => (
-                  <tr key={row.id}>
-                     {row.getVisibleCells().map((cell) => (
-                        <td
-                           key={cell.id}
-                           style={{
-                              textAlign: "left",
-                              paddingLeft: "20px",
-                              border: "1px solid black",
-                           }}
-                        >
-                           {flexRender(
-                              cell.column.columnDef.cell,
-                              cell.getContext()
-                           )}
-                        </td>
-                     ))}
-                  </tr>
-               ))}
-            </tbody>
-         </table>
-      </div>
-   );
-}
-
 type TableListProps = {
    tables: Table[];
    selectedTable: Table;
@@ -523,7 +317,7 @@ type Table = {
    name: string;
 };
 
-const useTableList = (projectId: string) => {
+export const useTableList = (projectId: string) => {
    const [tables, setTables] = useState<Table[]>([]);
    const { data, error } = useSWR(
       `${mockApiBuilder(projectId)}/data/all`,
@@ -545,39 +339,6 @@ const useTableList = (projectId: string) => {
    return { tables, error };
 };
 
-function TableEditorSidebar({ projectId, tableId, onChange }) {
-   const { tables, error } = useTableList(projectId);
-
-   return (
-      <div
-         className="sidebar"
-         style={{ padding: "30px", maxWidth: "220px", minWidth: "200px" }}
-      >
-         <h3>Tables</h3>
-         <ul className="table-list">
-            {error ? <div>Failed to load</div> : <></>}
-            {tables ? (
-               tables.map((table) => (
-                  <li
-                     key={table.id}
-                     className={`table-item ${
-                        table.id === tableId ? "selected" : ""
-                     }`}
-                     onClick={() =>
-                        onChange({ type: "select-table", tableId: table.id })
-                     }
-                  >
-                     {table.name}
-                  </li>
-               ))
-            ) : (
-               <div>Loading...</div>
-            )}
-         </ul>
-      </div>
-   );
-}
-
 export function TableEditor({ projectId, tableId }: TableEditorProps) {
    const [canva, canvaDispatch, config, configDispatch] = useTableEditor(
       projectId,
@@ -598,10 +359,8 @@ export function TableEditor({ projectId, tableId }: TableEditorProps) {
       }
 
       if (canva.isRowConfigOpen) {
-         console.log("set row config");
          setRowConfig(true);
       } else if (canva.isColumnConfigOpen) {
-         console.log("set column config");
          setColumnConfig(true);
       }
    }, [canva]);
@@ -632,17 +391,6 @@ export function TableEditor({ projectId, tableId }: TableEditorProps) {
             onClose={() => setColumnConfig(false)}
             dispatch={canvaDispatch}
          />
-         {/* Config menu for editing table */}
-         {/*<EditorConfigMenu
-			isOpen={config.isConfigOpen}
-			onClose={() => configDispatch({ type: "open-config" })}
-			onEdit={() => configDispatch({ type: "edit-config" })}
-			onAddColumn={() => configDispatch({ type: "add-column" })}
-			onAddRow={() => configDispatch({ type: "add-row" })}
-			onRemoveColumn={() => configDispatch({ type: "remove-column" })}
-			onRemoveRow={() => configDispatch({ type: "remove-row" })}
-			onSave={() => configDispatch({ type: "save" })}
-		 />*/}
       </div>
    );
 }
