+++ apps/web/app/[project-id]/data/page.tsx
@@ -1,13 +1,11 @@
 "use client";
 
 import { useParams, useRouter } from "next/navigation";
-import { TableList } from "./TableList";
+import { TableList } from "./components/TableList";
 
 export default function Page() {
    const params = useParams();
    const router = useRouter();
-   console.log("params", params);
-   // check if params contains 'project-id'
    if (params["project-id"]) {
       return (
          <>
